requirements.txt
absl-py==0.7.0
astor==0.7.1
backports.weakref==1.0.post1
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
google-api-core==1.7.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
h5py==2.9.0
httplib2==0.12.0
idna==2.8
joblib==0.13.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mock==2.0.0
numpy==1.16.0
pandas==0.23.4
pbr==5.1.1
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.19.1
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
